requirements.txt
pymupdf
python-docx
python-pptx
pandas
sentence-transformers
faiss-cpu
streamlit
unstructured
huggingface_hub
requests